Added boolean control

+++ main.go
@@ -22,7 +22,8 @@ type Entries struct {
 	Id   string `json:"id"`
 }
 
-var token = "Bearer sl.BHCO8jgBhjcpXkgTjzDGIr50l43YHZZndyfklZmZYIwah5wdSwRdH3jrJ_8kyFfKiP98Gc-yJf9lt3bwNtKpd_KZLmiluHgv3P6rr8QG-LI9dRXWfG40NMDcSR4YOLWde9YEZCAxXIGo"
+var delete = false
+var token = "Bearer " + "sl.BHCbUv37-DRLODPeBojZ99ui9O2bdnBND_s4ajjn2YL-ovTNxBvkqgwKwJaR9zZH1-zoEyImbxPnVZ44dtmnBr8vOatM1pJuyDRfi8xHr11blxMhHBP7ZCO0-hwWsrmE3UjVVcRuGEQD"
 
 func deleteFile(path string) {
 	url := "https://api.dropboxapi.com/2/files/delete_v2"
@@ -56,7 +57,6 @@ func getData() {
 	req, err := http.NewRequest("POST", url, bytes.NewBuffer(jsonStr))
 
 	req.Header.Add("Authorization", token)
-	//request.Header.Add("cursor", response.Cursor)
 	req.Header.Set("Content-Type", "application/json")
 	res, _ := http.DefaultClient.Do(req)
 
@@ -69,18 +69,19 @@ func getData() {
 	var response Response
 	json.Unmarshal(body, &response)
 
-	//fmt.Println(response.Entries)
 	for e, f := range response.Entries {
-		fmt.Println("File numero:", e)
-		fmt.Println("Tag:", f.Tag)
-		fmt.Println("Name:", f.Name)
 		if string(f.Name[0]) == "~" {
-			deleteFile(f.Path)
+			if delete == true {
+				deleteFile(f.Path)
+			}
+			fmt.Println("File numero:", e)
+			fmt.Println("Tag:", f.Tag)
+			fmt.Println("Name:", f.Name)
 			fmt.Println("Path:", f.Path)
+			fmt.Println("Id:", f.Id)
+			fmt.Println("--------------------------------")
 		}
-		fmt.Println("Path:", f.Path)
-		fmt.Println("Id:", f.Id)
-		fmt.Println("--------------------------------")
+
 	}
 
 }
